+++ src/components/index.js
@@ -2,3 +2,4 @@ export * from './utils/makeArray';
 export * from './utils/makeDelaySearch';
 export * from './utils/search';
 export * from './utils/numberFormat';
+export * from './utils/numberDecimal';