+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+##### 1.5.0
+
+| name              | Type               | Version              |
+| :---------------- | :----------------- |:---------------------|
+| makeArray         | Boolean            | 1.0.0 
+| makeDelaySearch   | Boolean            | 1.0.0   
+| search            | Boolean            | 1.0.0   
+| numberFormat      | Boolean            | 1.0.0   
+| numberDecimal     | Boolean            | 1.0.0   
+
 ##### 1.0.0
 
 - Initial bnv-utils